requirements.txt
twitchio==2.10.0
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3 